requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=3.5.0=py38hecd8cb5_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38hca72f7f_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py38hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py38h9ed2024_1003
ca-certificates=2022.07.19=hecd8cb5_0
certifi=2022.9.14=py38hecd8cb5_0
cffi=1.15.1=py38hc55c11b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py38hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py38he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.69.1=h8346a28_1
gst-plugins-base=1.14.0=h4180768_2
gstreamer=1.14.0=h0fc69c2_2
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.15.2=py38hecd8cb5_0
ipython=8.4.0=py38hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=hca72f7f_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py38hecd8cb5_0
jupyter=1.0.0=py38hecd8cb5_8
jupyter_client=7.3.5=py38hecd8cb5_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py38hecd8cb5_0
jupyter_server=1.18.1=py38hecd8cb5_0
jupyterlab=3.4.4=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py38hecd8cb5_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py38he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libclang=12.0.0=default_hbc2896b_2
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libffi=3.3=hb1e8313_2
libiconv=1.16=hca72f7f_2
libllvm12=12.0.0=h9b2ccf5_3
libpng=1.6.37=ha441bb4_0
libpq=12.9=h1c9f633_3
libsodium=1.0.18=h1de35cc_0
libtiff=4.4.0=h2ef1027_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
libxslt=1.1.35=h5b33f42_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.3=h23ab428_1
markupsafe=2.1.1=py38hca72f7f_0
matplotlib=3.5.2=py38hecd8cb5_0
matplotlib-base=3.5.2=py38hfb0c5b7_0
matplotlib-inline=0.1.6=py38hecd8cb5_0
mistune=0.8.4=py38h1de35cc_1001
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
munkres=1.1.4=py_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py38hecd8cb5_0
nbconvert=6.4.4=py38hecd8cb5_0
nbformat=5.5.0=py38hecd8cb5_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.5=py38hecd8cb5_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
notebook=6.4.12=py38hecd8cb5_0
nspr=4.33=he9d5cce_0
nss=3.74=h47edf6a_0
numpy=1.23.1=py38h2e5f0a9_0
numpy-base=1.23.1=py38h3b1a694_0
openssl=1.1.1q=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hde71d04_1
pip=22.1.2=py38hecd8cb5_0
ply=3.11=py38_0
prometheus_client=0.14.1=py38hecd8cb5_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
psutil=5.9.0=py38hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py38hecd8cb5_0
pyqt=5.15.7=py38he9d5cce_0
pyqt5-sip=12.11.0=py38he9d5cce_0
pyrsistent=0.18.0=py38hca72f7f_0
pysocks=1.7.1=py38_1
python=3.8.13=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38hecd8cb5_0
pytorch=1.9.0=py3.8_0
pytz=2022.1=py38hecd8cb5_0
pyzmq=23.2.0=py38he9d5cce_0
qt-main=5.15.2=h719ae48_7
qt-webengine=5.15.9=h90a370e_4
qtconsole=5.3.2=py38hecd8cb5_0
qtpy=2.2.0=py38hecd8cb5_0
qtwebkit=5.212=h24dc246_4
readline=8.1.2=hca72f7f_1
requests=2.28.1=py38hecd8cb5_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py38hecd8cb5_0
sip=6.6.2=py38he9d5cce_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py38hecd8cb5_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.39.2=h707629a_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py38hecd8cb5_0
testpath=0.6.0=py38hecd8cb5_0
tk=8.6.12=h5d9f67b_0
toml=0.10.2=pyhd3eb1b0_0
torchtext=0.10.0=py38
tornado=6.2=py38hca72f7f_0
tqdm=4.64.0=py38hecd8cb5_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py38hecd8cb5_0
typing_extensions=4.3.0=py38hecd8cb5_0
urllib3=1.26.11=py38hecd8cb5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38hecd8cb5_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38hecd8cb5_0
xz=5.2.5=hca72f7f_1
zeromq=4.3.4=h23ab428_0
zipp=3.8.0=py38hecd8cb5_0
zlib=1.2.12=h4dc903c_3
zstd=1.5.2=hcb37349_0
